pyproject.toml
[tool.poetry]
name = "ls-cleaner"
version = "0.1.0"
description = ""
authors = ["Sultan Zakariya <sultan@epoch8.co>"]
readme = "README.md"
packages = [{include = "ls_cleaner"}]

[tool.poetry.dependencies]
python = "^3.9"
label-studio-sdk = "^0.0.28"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
